requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
neomodel==5.3.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
